LayoutTests/imported/w3c/web-platform-tests/IndexedDB/ready-state-destroyed-execution-context-expected.txt


FAIL readyState accessor is valid after execution context is destroyed assert_equals: expected "done" but got "pending"
